Started with class based views with editing paper and listing papers.
Using only the main View class. So far generated one base view and inherited it to reuse the get and post methods but created a new get_context_data method. It may take a few classes to be a able to generate efficient code. For now just convert all the code and then generate an inheritance tree.