Basic web scraping from IEEE website and inserting into database created.
The problem now is that large writes to the database are agonizingly slow because of the save on every entry. To speed it up I need to sue a bulk write to the database but with foreign key and many-to-many relationships that is not possible. This means that bulk writes will need to be done manually through raw SQL commands. A foreign key or many-to-many relationship is another table. This table will need to be created manually and fitted into the Django ORM.