Changed the data structure for Paper to be independent and Author and Journal to add Paper as M2M.
To begin with, the fundamental component in the database is the Paper model as it is the papers that are extracted either from the web or from a BibTex file. So to speed up the initial database population, it is advisable that Paper be completely independent so that bulk inserts of Paper through bulk_create is possible. Author and Journal now add Paper as ManyToMany. Author adding paper seems intuitive while journal not so. But if we set aside the assumption that a paper can be published in only one place, it is possible. This results in very fast database population within seconds even when the number of papers is in hundreds which is a huge improvement over the 20 minutes it took for the database write before. The next step will be let the user add papers to the author and to journals.