Bringing back most of the old editing features for papers and authors but with bulk_create to speed up database writes.
The paper can be edited as before with the additional options of editing the authors so as to make it possible to manually add an author. Papers can be added to authors. The next step is to be able to show additional entries of the same author to delete additions. Furthermore, the same needs to be done to Journals.