Started with the journal section with a basic browse publishers link working.
The next step is to be able to edit the journal entries and add papers to the journals. Repititions to journals are not added unlike authors as journals are less diverse than authors. It is left completely to the user to edit and delete authors since it is possible that two authors with the same name exist. However, two publishers with the same name is unlikely.